refactor: clean up monaco editor integration
- Move monacoLinter from utils to services folder
- Remove duplicate monaco-linter.js file
- Remove all debug console.log statements
- Simplify worker configuration for Python-only usage
- Update imports to use services folder
